+++ d60.Cirqus.Tests/Contracts/Views/Models/ViewInstanceDeletion/SomethingHappened.cs
@@ -1,0 +1,9 @@
+﻿using d60.Cirqus.Events;
+
+namespace d60.Cirqus.Tests.Contracts.Views.Models.ViewInstanceDeletion
+{
+    public class SomethingHappened : DomainEvent<Root>
+    {
+        
+    }
+}