+++ d60.Cirqus.Tests/Contracts/Views/UpdatedEvent.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using d60.Cirqus.Tests.Contracts.Views.Factories;
 using d60.Cirqus.Tests.Contracts.Views.Models.UpdatedEvent;
 using NUnit.Framework;
@@ -50,9 +49,10 @@ namespace d60.Cirqus.Tests.Contracts.Views
             _context.Save("id2", new Event());
 
             // assert
+            Assert.That(registeredUpdates["id1"], Is.GreaterThanOrEqualTo(3));
+            Assert.That(registeredUpdates["id2"], Is.GreaterThanOrEqualTo(1));
+
             Assert.That(registeredUpdates.Count, Is.EqualTo(2));
-            Assert.That(registeredUpdates["id1"], Is.EqualTo(3));
-            Assert.That(registeredUpdates["id2"], Is.EqualTo(1));
         }
     }
 }