+++ d60.Cirqus.Tests/Views/NewViewManager/TestViewManagerEventDispatcher.cs
@@ -33,7 +33,7 @@ namespace d60.Cirqus.Tests.Views.NewViewManager
             CirqusLoggerFactory.Current = new ConsoleLoggerFactory(minLevel: Logger.Level.Debug);
 
             _mongoDatabase = MongoHelper.InitializeTestDatabase();
-            
+
             _commandProcessor = CommandProcessor.With()
                 .Logging(l => l.UseConsole(minLevel: Logger.Level.Warn))
                 .EventStore(e => e.UseMongoDb(_mongoDatabase, "Events"))
@@ -144,7 +144,7 @@ namespace d60.Cirqus.Tests.Views.NewViewManager
         {
             var allPotatoesView = new MongoDbViewManager<AllPotatoesView>(_mongoDatabase);
             var potatoTimeToBeConsumedView = new MongoDbViewManager<PotatoTimeToBeConsumedView>(_mongoDatabase);
-            
+
             _dispatcher.AddViewManager(allPotatoesView);
             _dispatcher.AddViewManager(potatoTimeToBeConsumedView);